+++ ScopedVolumeWheel/VolumeDisplay.h
@@ -15,8 +15,8 @@ public:
 private:
 	HWND hWnd = 0;
 	float level = 0;
-	long long duration = 0.5 * 1e9;
-	long long fadeTime = 0.5 * 1e9;
+	long long duration = static_cast<long long>(0.5 * 1e9);
+	long long fadeTime = static_cast<long long>(0.5 * 1e9);
 	bool visible = false;
 	std::chrono::steady_clock clock;
 	std::chrono::steady_clock::time_point startTime;
